feat: Add a button to create new steps in the EditScreen

+++ app/src/main/java/com/mrdarip/stepy/data/local/entities/StepEntity.kt
@@ -21,5 +21,5 @@ data class StepEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val name: String,
     val position: Int,
-    val taskId: Int
+    val taskId: Long
 )
