+++ app/src/main/java/com/mrdarip/stepy/domain/model/Step.kt
@@ -4,5 +4,5 @@ data class Step(
     val id: Long = 0,
     val name: String,
     val position: Int,
-    val taskId: Int
+    val taskId: Long
 )
